refactor: extract hook
